Fix missing eigenvalue difference calculation
At present, `deig` actually corresponds to the bulk_index eigenvalues, rather than the difference in eigenvalues between the bulk and defect.
Tested and seems to work as expected. (I noticed this from sanity testing with a manual calculation of Wif using the WSWQ, dQ and eigenvalues, which was giving me different results at first, but this fixed it).